requirements.txt
amqp==5.1.1
asgiref==3.8.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2024.7.4
cffi==1.15.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
daphne==4.0.0
cryptography==42.0.8
decorator==5.1.1
Django==5.0.7
django-cors-headers==3.13.0
django-environ==0.9.0
django-eventstream==4.5.1
django-extensions==3.2.1
django-filter==23.5
django-redis==5.2.0
django-safedelete==1.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-websocket-redis==0.6.0
docopt==0.6.2
drf-api-tracking==1.8.0
drf-extensions==0.7.1
drf-spectacular==0.27.2
gssapi==1.8.2
gunicorn==22.0.0
hdfs==2.7.0
hiredis==2.2.2
idna==3.7
importlib-metadata==5.2.0
inflection==0.5.1
influxdb-client==1.36.1
itypes==1.2.0
Jinja2==3.1.4
kerberos==1.3.1
kombu==5.2.4
krb5==0.4.1
Markdown==3.4.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==22.0
prompt-toolkit==3.0.36
psycopg[binary]==3.1.16
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.4.0
PyJWT==2.6.0
pyOpenSSL==24.1.0
pyspnego==0.7.0
python-json-logger==2.0.7
pytest==7.4.0
PyYAML==6.0
redis==4.6.0
requests==2.32.3
requests-kerberos==0.14.0
ruamel.yaml==0.17.21
ruff==0.5.0
setuptools==70.0.0
simplejson==3.18.0
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==1.26.19
vine==5.0.0
wcwidth==0.2.5
zipp==3.19.1
